test: fixing more tests
