+++ packages/next/src/data/hooks/__tests__/useAppSettings.tsx
@@ -1,15 +1,22 @@
 import { setHeadstartWPConfig, type AppEntity, type EndpointParams } from '@headstartwp/core';
 import { expectTypeOf } from 'expect-type';
 import { renderHook } from '@testing-library/react';
+import * as React from 'react';
+import { SettingsProvider } from '@headstartwp/core/react';
 import { useAppSettings } from '../useAppSettings';
+
+const config = {
+	sourceUrl: 'https://js1.10up.com',
+	useWordPressPlugin: true,
+};
 
 describe('useAppSettings types', () => {
 	beforeAll(() => {
-		setHeadstartWPConfig({
-			sourceUrl: 'https://js1.10up.com',
-			useWordPressPlugin: true,
-		});
+		setHeadstartWPConfig(config);
 	});
+	const wrapper = ({ children }) => {
+		return <SettingsProvider settings={config}>{children}</SettingsProvider>;
+	};
 
 	it('allows overriding types', () => {
 		interface MyAppEntity extends AppEntity {
@@ -20,8 +27,9 @@ describe('useAppSettings types', () => {
 			includeCustomSettings: boolean;
 		}
 
-		const { result } = renderHook(() =>
-			useAppSettings<MyAppEntity, Params>({ includeCustomSettings: true }),
+		const { result } = renderHook(
+			() => useAppSettings<MyAppEntity, Params>({ includeCustomSettings: true }),
+			{ wrapper },
 		);
 
 		expectTypeOf(result.current.data).toMatchTypeOf<
