+++ packages/next/src/data/hooks/__tests__/useTerms.tsx
@@ -4,16 +4,28 @@ import {
 	type TermEntity,
 } from '@headstartwp/core';
 import { renderHook } from '@testing-library/react';
+import * as React from 'react';
 import { expectTypeOf } from 'expect-type';
+import { DataFetchingProvider, SettingsProvider } from '@headstartwp/core/react';
 import { useTerms } from '../useTerms';
+
+const config = {
+	sourceUrl: 'https://js1.10up.com',
+	useWordPressPlugin: true,
+};
 
 describe('useTerms types', () => {
 	beforeAll(() => {
-		setHeadstartWPConfig({
-			sourceUrl: 'https://js1.10up.com',
-			useWordPressPlugin: true,
-		});
+		setHeadstartWPConfig(config);
 	});
+
+	const wrapper = ({ children }) => {
+		return (
+			<DataFetchingProvider swrConfig={{ provider: () => new Map() }} data={{}}>
+				<SettingsProvider settings={config}>{children}</SettingsProvider>
+			</DataFetchingProvider>
+		);
+	};
 
 	it('allows overriding types', () => {
 		interface Genre extends TermEntity {
@@ -24,7 +36,9 @@ describe('useTerms types', () => {
 			editor: string;
 		}
 
-		const { result } = renderHook(() => useTerms<Genre, GenreParams>({ editor: 'sdasd' }));
+		const { result } = renderHook(() => useTerms<Genre, GenreParams>({ editor: 'sdasd' }), {
+			wrapper,
+		});
 
 		expectTypeOf(result.current.data?.terms).toMatchTypeOf<
 			| Array<{
